feat: Transactions for voting
When voting in batches, we need to do it as transactions as we else can have some big problems where only some of the candidate ranks gets commited to the database. This commit also includes some validations for batch voting, so now we are checking if candidate id is valid with the use of Laravel validation.

Commit includes:
- Transaction for batch voting
- Validation in batch voting
- Refactoring of batch voting

Resolves: #94